Move all script to index.js

+++ index.js
@@ -1,22 +1,58 @@
+#!/usr/bin/env node
 'use strict'
-const fs = require('fs')
-const shelljs = require('shelljs')
-const url = 'https://github.com/milligram/milligram-starter.git'
 
-function init (argv) {
-  if (!shelljs.which('git')) {
-    console.error('[error] Sorry, this script requires git')
-    process.exit(1)
-  }
-  const directory = argv[1] || 'milligram_app'
+const fs = require('fs')
+const argv = require('rasper')(process.argv.slice(2))
+const shelljs = require('shelljs')
+const pkg = require('./package.json')
+
+if (argv.v || argv.version) {
+  console.log(pkg.version)
+  process.exit(0)
+}
+
+if (argv.h || argv.help) {
+  console.log(`
+  $ milligram --help
+
+    Usage:
+
+      $ milligram <command> <directory> [<options>]
+
+    Commands:
+
+      i, init                 Initialize a Milligram project
+
+    Options:
+
+      -h, --help              Display help information
+      -v, --version           Output version
+
+    Examples:
+
+      $ milligram init new_app
+  `)
+  process.exit(0)
+}
+
+if (argv._[0]) {
+  const directory = argv._[0] || 'milligram_app'
+
   if (fs.existsSync(directory)) {
     console.warn('[warning] This directory already exists.')
     process.exit(1)
   }
-  shelljs
-    .exec(`git clone ${url} ${directory}`)
-    .exec(`rm -rf ${directory}/.* 2> /dev/null`)
-  process.exit(1)
-}
 
-exports.init = init
+  if (!shelljs.which('git')) {
+    console.error('[error] Sorry, this script requires git')
+    process.exit(1)
+  }
+
+  shelljs
+    .exec(
+      `git clone https://github.com/milligram/milligram-starter.git ${directory}`,
+    )
+    .exec(`rm -rf ${directory}/.* 2> /dev/null`)
+
+  process.exit(0)
+}
